H-GPT/requirements_deploy.txt
tensorboard
pytorch_lightning
accelerate
torchmetrics
omegaconf
shortuuid
chumpy
transformers
diffusers
einops
wandb
rich
matplotlib

# for visualization
smplx
trimesh
joblib==1.2.0
h5py
scikit-image
spacy
ftfy
more-itertools
natsort
pyrender
moviepy
librosa
triangle
easydict
scipy
gradio==3.50.2
numpy==1.23.1
shapely==2.0.6
pybullet==3.2.6
opencv-python
bert_score==0.3.13
peft==0.14.0